Added initial support for equirectangular maps in opengl.
- Added support for compute shaders in opengl.

+++ Humble2/src/Humble2/Renderer/ComputePassRenderer.h
@@ -12,7 +12,7 @@ namespace HBL2
 	{
 		Handle<Shader> Shader;
 		Handle<BindGroup> BindGroup;
-		glm::vec3 ThreadGroupCount{};
+		glm::uvec3 ThreadGroupCount{};
 	};
 
 	class HBL2_API ComputePassRenderer
